requirement.txt
steemit
slacker